requirements.txt
pydub==0.25.1
moviepy==1.0.3
Pillow>=10.0.0
python-dotenv==1.0.1
